loud CR
