+++ packages/vscode-extension/src/project/metro.ts
@@ -10,6 +10,7 @@ import { shouldUseExpoCLI } from "../utilities/expoCli";
 import { Devtools } from "./devtools";
 import { getLaunchConfiguration } from "../utilities/launchConfiguration";
 import { EXPO_GO_BUNDLE_ID, EXPO_GO_PACKAGE_NAME } from "../builders/expoGo";
+import { evaluateJsFromCdp } from "../utilities/evaluateJsFromCdp";
 
 export interface MetroDelegate {
   onBundleError(): void;
@@ -297,74 +298,11 @@ export class Metro implements Disposable {
     source: string,
     timeoutMs: number = 2000
   ): Promise<string | undefined> {
-    const REQUEST_ID = 0;
-
     const wsUrl = new URL(webSocketDebuggerUrl);
     wsUrl.hostname = "localhost";
     wsUrl.port = String(this._port);
 
-    return new Promise((resolve, reject) => {
-      let settled = false;
-      const ws = new WebSocket(wsUrl);
-
-      const timeoutHandle = setTimeout(() => {
-        Logger.debug(`[evaluateJsFromCdpAsync] Request timeout from ${wsUrl.toString()}`);
-        reject(new Error("Request timeout"));
-        settled = true;
-        ws.close();
-      }, timeoutMs);
-
-      const settleConnection = () => {
-        settled = true;
-        clearTimeout(timeoutHandle);
-        ws.close();
-      };
-
-      ws.on("open", () => {
-        ws.send(
-          JSON.stringify({
-            id: REQUEST_ID,
-            method: "Runtime.evaluate",
-            params: { expression: source },
-          })
-        );
-      });
-
-      ws.on("error", (e) => {
-        Logger.debug(`[evaluateJsFromCdpAsync] Failed to connect ${wsUrl.toString()}`, e);
-        reject(e);
-        settleConnection();
-      });
-
-      ws.on("close", () => {
-        if (!settled) {
-          reject(new Error("WebSocket closed before response was received."));
-          clearTimeout(timeoutHandle);
-        }
-      });
-
-      ws.on("message", (data) => {
-        Logger.debug(
-          `[evaluateJsFromCdpAsync] message received from ${wsUrl.toString()}: ${data.toString()}`
-        );
-        try {
-          const response = JSON.parse(data.toString());
-          if (response.id === REQUEST_ID) {
-            if (response.error) {
-              reject(new Error(response.error.message));
-            } else if (response.result.result.type === "string") {
-              resolve(response.result.result.value);
-            } else {
-              resolve(undefined);
-            }
-            settleConnection();
-          }
-        } catch (e) {
-          reject(e);
-          settleConnection();
-        }
-      });
-    });
+    return evaluateJsFromCdp(wsUrl.toString(), source);
   }
 
   public async reload() {
